chore(remove): temp removed hold to count feat to be added later
